refactor(rte-table): make the TableDragHandles public surface explicit
The folder was ten flat peers with nothing to say which were importable.
Exactly one of them is: TipTapEditor/components.tsx and one test consume
`TableDragHandles`, and nothing else reaches in.

Structure now states that:

  TableDragHandles/
    index.ts          <- public surface, one export
    TableDragHandles.tsx
    internal/         <- private to this component

Also folds utils/tableEditorGeometry.ts in as internal/coordinates.ts. It was
sitting in the shared feature surface but had no consumer outside this folder,
so it was reachable by anything while being private in practice.
utils/tableEditorChrome.ts stays put — TableCaption/TableNodeView.tsx uses
TABLE_ADD_CHROME_PX, so it is genuinely shared.

Kept as convention rather than a lint rule; all moves are git renames so
history follows the files.
