+++ apps/studio/src/features/editing-experience/components/TableDragHandles/index.ts
@@ -1,0 +1,4 @@
+// The public surface of this folder. Everything under `internal/` is private
+// to TableDragHandles and should not be imported from outside.
+export { TableDragHandles } from "./TableDragHandles"
+export type { TableDragHandlesProps } from "./TableDragHandles"
