+++ apps/studio/src/features/editing-experience/components/TableDragHandles/internal/coordinates.ts
